report/src/events/listeners/user-created-listener.ts
import { UserCreatedEvent, Subjects, Listener } from "@manthan-practice/common";
import { Message } from "node-nats-streaming";
import { queueGroupName } from "./queue-group-name";
import { User } from "../../models/user";

export class UserCreatedListener extends Listener<UserCreatedEvent> {
  readonly subject = Subjects.UserCreated;
  queueGroupName = queueGroupName;

  async onMessage(data: UserCreatedEvent["data"], msg: Message) {
    const user = await User.build(data);
    await user.save();
    msg.ack();
  }
}
